cassandra_sample/requirements.txt
cassandra-driver==3.22.0